requirements.txt
Django==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.10.1
pytz==2019.1
sqlparse==0.3.0
